Rename local aiosqlite and jwt helpers
