+++ services/sponsorship_service.py
@@ -1,5 +1,8 @@
 # File: backend/services/sponsorship_service.py
-import aiosqlite
+try:  # pragma: no cover - prefer local stub if available
+    import aiosqlite_local as aiosqlite
+except ModuleNotFoundError:  # pragma: no cover - fallback to package
+    import aiosqlite  # type: ignore
 from typing import Optional, Dict, Any, List
 from datetime import date
